+++ src/berconpy/parser.py
@@ -158,7 +158,12 @@ def _get_pattern_kwargs(m: re.Match) -> dict:
 
 def is_expected_message(message: str) -> bool:
     """Determines if a server message is expected."""
-    return message == "Connected to BE Master"
+    return message in (
+        "Ban check timed out, no response from BE Master",
+        "Connected to BE Master",
+        "Disconnected from BE Master",
+        "Failed to resolve BE Master DNS name(s)",
+    )
 
 
 def parse_admins(response: str) -> Iterator[ParsedAdmin]: